requirements.txt
python-dateutil==2.6.1
pandas
snowflake-connector-python
snowflake-sqlalchemy
